+++ web/src/hooks/useWhitelistedTokens.ts
@@ -5,31 +5,10 @@ import {
   type QueryClient,
 } from "@tanstack/react-query";
 import { gatewayAddresses } from "@vetro-protocol/gateway";
-import { fetchWhitelistedTokens } from "fetchers/fetchWhitelistedTokens";
-import type { Address, Client } from "viem";
+import type { Client } from "viem";
 
 import { useEthereumClient } from "./useEthereumClient";
-
-export const whitelistedTokensByGatewayOptions = ({
-  client,
-  gatewayAddress,
-  queryClient,
-}: {
-  client: Client | undefined;
-  gatewayAddress: Address;
-  queryClient: QueryClient;
-}) =>
-  queryOptions({
-    enabled: !!client && !!client.chain,
-    queryFn: () =>
-      fetchWhitelistedTokens({
-        client: client!,
-        gatewayAddress,
-        queryClient,
-      }),
-    queryKey: ["whitelisted-tokens", client?.chain?.id, gatewayAddress],
-    staleTime: Infinity,
-  });
+import { whitelistedTokensByGatewayOptions } from "./useWhitelistedTokensByGateway";
 
 const whitelistedTokensOptions = ({
   client,
@@ -56,17 +35,9 @@ const whitelistedTokensOptions = ({
     staleTime: Infinity,
   });
 
-export const useWhitelistedTokens = function (gatewayAddress?: Address) {
+export const useWhitelistedTokens = function () {
   const client = useEthereumClient();
   const queryClient = useQueryClient();
 
-  return useQuery(
-    gatewayAddress
-      ? whitelistedTokensByGatewayOptions({
-          client,
-          gatewayAddress,
-          queryClient,
-        })
-      : whitelistedTokensOptions({ client, queryClient }),
-  );
+  return useQuery(whitelistedTokensOptions({ client, queryClient }));
 };
