+++ web/src/pages/analytics/components/yieldCard.tsx
@@ -1,5 +1,5 @@
 import { useAnalyticsTreasury } from "hooks/useAnalyticsTreasury";
-import { useWhitelistedTokens } from "hooks/useWhitelistedTokens";
+import { useWhitelistedTokensByGateway } from "hooks/useWhitelistedTokensByGateway";
 import { useTranslation } from "react-i18next";
 import type { TokenWithGateway } from "types";
 
@@ -16,7 +16,7 @@ type Props = {
 export const YieldCard = function ({ peggedToken, peggedTokenError }: Props) {
   const { t } = useTranslation();
   const { data: whitelistedTokens, isError: isWhitelistedTokensError } =
-    useWhitelistedTokens(peggedToken?.gatewayAddress);
+    useWhitelistedTokensByGateway(peggedToken?.gatewayAddress);
   const {
     data: treasury,
     isError: isTreasuryError,
